// apps/api/src/resolvers-types.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
};

export type Message = {
  __typename?: 'Message';
  content: Scalars['String']['output'];
  createdAt: Scalars['String']['output'];
  id: Scalars['ID']['output'];
  readAt?: Maybe<Scalars['String']['output']>;
  sentBy: SentBy;
};

export type Mutation = {
  __typename?: 'Mutation';
  markMessageAsRead?: Maybe<Message>;
  sendMessage: Message;
};


export type MutationMarkMessageAsReadArgs = {
  id: Scalars['ID']['input'];
};


export type MutationSendMessageArgs = {
  content: Scalars['String']['input'];
  sentBy: SentBy;
};

export type Query = {
  __typename?: 'Query';
  getMessages: Array<Message>;
};

export enum SentBy {
  System = 'SYSTEM',
  User = 'USER'
}